Refactor DSL options into a separate module
