+++ lib/avro/builder/errors.rb
@@ -40,12 +40,8 @@ module Avro
 
     class UnsupportedBlockAttributeError < StandardError
       def initialize(attribute:, type:, field: nil)
-        target = if field
-                   "field '#{field}' of type :#{type}"
-                 else
-                   "type :#{type}"
-                 end
-        super("'#{attribute}' must be set directly using an option on #{target}")
+        super("'#{attribute}' must be set directly using an option on "\
+              "field '#{field}' of type :#{type}")
       end
     end
   end
